--- a/dataprep/eda/correlation/compute/__init__.py
+++ b/dataprep/eda/correlation/compute/__init__.py
@@ -8,6 +8,8 @@ from ...intermediate import Intermediate
 from .bivariate import _calc_bivariate
 from .nullivariate import _calc_nullivariate
 from .univariate import _calc_univariate
+from ...dtypes import NUMERICAL_DTYPES
+from ...utils import to_dask
 
 __all__ = ["compute_correlation"]
 
@@ -34,8 +36,10 @@ def compute_correlation(
     k
         Choose top-k element
     """
-
-    df = DataArray(df).select_num_columns()
+    if x is not None and y is not None:
+        df = to_dask(df.select_dtypes(NUMERICAL_DTYPES))
+    else:
+        df = DataArray(df).select_num_columns()
 
     if x is None and y is None:  # pylint: disable=no-else-return
         return _calc_nullivariate(df, value_range=value_range, k=k)