# Description

Optimized plot_missing(df, x) and plot_correlation(df, x, y). The main change was using dataframe API rather than array. Closes #319 

# How Has This Been Tested?

manually

# Snapshots:

`plot_missing(df, x)` example increase 47s ->6s

![Screen Shot 2020-09-18 at 2 45 41 PM](https://user-images.githubusercontent.com/25874021/93652385-ae864800-f9c9-11ea-9bae-f0013e6d31d9.png)

![Screen Shot 2020-09-18 at 4 11 35 PM](https://user-images.githubusercontent.com/25874021/93652394-b8a84680-f9c9-11ea-8da4-506c09dfc4f0.png)


`plot_correlation(df, x, y)` example increase 100s -> ~1s

![Screen Shot 2020-09-18 at 12 37 36 PM](https://user-images.githubusercontent.com/25874021/93652212-2011c680-f9c9-11ea-8ef5-27fd84721edd.png)

![Screen Shot 2020-09-18 at 12 31 21 PM](https://user-images.githubusercontent.com/25874021/93652229-3586f080-f9c9-11ea-9708-81717cf9e604.png)


# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have already squashed the commits and make the commit message conform to the project standard.
- [x] I have already marked the commit with "BREAKING CHANGE" or "Fixes #" if needed.
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules